+++ material/src/material.ts
@@ -1,6 +1,1 @@
-export * from './modules/dialogPopup.module';
-export * from './components/basicDialogPopup/types';
-export * from './plugins/popup/components';
-export * from './plugins/popup';
-export * from './components/basicDialogPopup';
-export * from './directives/types';
+export * from './services/dialogInternalServerErrorRenderer.service';